use a friendly sort for all viewslices, persist collapse on summaries
Because tabs can change their sort location when they navigate, we need
to remove and re-add the binding.  This would be destructive to collapse
state if we didn't store it on the summary object.  It's not entirely
hygienic; we might be better off using an idspace persistence mechanism.
But it's workable for the time being.